chore: :fire: restructured remaining components
