+++ src/features/Login/Login.module.scss
@@ -27,3 +27,7 @@
 .footer {
   width: 100%;
 }
+
+.action {
+  margin-top: rem(1.5);
+}
