--- src/features/dashboard/components/Table/index.ts
@@ -1,4 +1,0 @@
-import Table from './table';
-
-
-export default Table;
