+++ src/features/dashboard/components/api-token-table/api-token-table.tsx
@@ -8,8 +8,8 @@ import Spinner from '@site/src/components/Spinner';
 import useApiToken from '@site/src/hooks/useApiToken';
 import useDeviceType from '@site/src/hooks/useDeviceType';
 import AccountSwitcher from '@site/src/components/AccountSwitcher';
-import ScopesCell from '../Table/CellScopes';
-import Table from '../Table';
+import ScopesCell from '../common-table/cell-scopes';
+import Table from '../common-table';
 import ApiTokenCell from './cell-token';
 import ApiLastUsedCell from './cell-last-used';
 import TokenActionsCell from './cell-delete-token';
