+++ src/features/dashboard/components/common-table/cell-copy-text.module.scss
