+++ src/features/dashboard/components/common-table/cell-scopes.module.scss
