+++ src/features/dashboard/components/common-table/cell-scopes.tsx
