+++ src/features/dashboard/components/common-table/common-table.scss
