+++ src/features/dashboard/components/common-table/index.ts
@@ -1,0 +1,3 @@
+import CommonTable from './common-table';
+
+export default CommonTable;
