+++ src/features/dashboard/components/token-name-restrictions/index.ts
@@ -1,0 +1,3 @@
+import TokenNameRestrictions from './token-name-restrictions';
+
+export default TokenNameRestrictions;
