+++ src/features/dashboard/components/token-name-restrictions/token-name-restrictions.module.scss
