+++ src/features/dashboard/components/token-register/index.ts
@@ -1,0 +1,3 @@
+import TokenRegister from './token-register';
+
+export default TokenRegister;
