+++ src/features/dashboard/components/token-register/token-register.scss
