+++ src/features/dashboard/manage-apps/index.ts
@@ -1,0 +1,3 @@
+import ManageApps from './manage-apps';
+
+export default ManageApps;
