+++ src/features/dashboard/register-tokens/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import TokenRegister from '../components/TokenRegister';
+import TokenRegister from '../components/token-register';
 
 const TokenRegistration: React.FC = () => {
   return <TokenRegister />;
