+++ src/pages/dashboard.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Layout from '@theme/Layout';
 import { AppManager } from '@site/src/features/dashboard';
+
 export default function AppRegistrationPage() {
   return (
     <Layout title='Dashboard' description='Manage your apps and api tokens'>
